Renamed the util functions
