+++ template/src/AppWrapper.tsx
@@ -18,7 +18,7 @@ import {SessionProvider} from './components/SessionContext';
 import {ImageBackground, SafeAreaView, StatusBar, Platform} from 'react-native';
 import ColorConfigure from './components/ColorConfigure';
 import Toast from '../react-native-toast-message';
-import useToastConfig from './subComponents/useToastConfig';
+import ToastConfig from './subComponents/ToastConfig';
 import {isValidReactComponent} from './utils/common';
 import DimensionProvider from './components/dimension/DimensionProvider';
 import Error from './components/common/Error';
@@ -37,7 +37,6 @@ const AppWrapper = (props: AppWrapperProps) => {
     }
     return React.Fragment;
   });
-  const toastConfig = useToastConfig();
 
   return (
     <AppRoot>
@@ -49,7 +48,7 @@ const AppWrapper = (props: AppWrapperProps) => {
           // @ts-ignore textAlign not supported by TS definitions but is applied to web regardless
           style={[{flex: 1}, Platform.select({web: {textAlign: 'left'}})]}>
           <StatusBar hidden={true} />
-          <Toast ref={(ref) => Toast.setRef(ref)} config={toastConfig} />
+          <Toast ref={(ref) => Toast.setRef(ref)} config={ToastConfig} />
           <StorageProvider>
             <GraphQLProvider>
               <Router>
