+++ template/src/atoms/TextInput.tsx
@@ -11,14 +11,13 @@
 */
 import React, {useContext} from 'react';
 import {TextInputProps, StyleSheet, TextInput, Platform} from 'react-native';
-import {useIsWebInternal} from '../utils/common';
+import {isWebInternal} from '../utils/common';
 import {textInput} from '../../theme.json';
 import ColorContext from '../components/ColorContext';
 
 const PrimaryButton = (props: TextInputProps) => {
   const {primaryColor} = useContext(ColorContext);
   const {style, ...otherProps} = props;
-  const styles = useStyles();
   return (
     <TextInput
       style={[
@@ -37,16 +36,13 @@ const PrimaryButton = (props: TextInputProps) => {
 
 export default PrimaryButton;
 
-const useStyles = () => {
-  const isWebInternal = useIsWebInternal();
-  return StyleSheet.create({
-    textInput,
-    // @ts-ignore
-    noOutline: isWebInternal() ? {outlineStyle: 'none'} : {},
-    textWrapFix: Platform.select({
-      ios: {
-        paddingVertical: 5,
-      },
-    }),
-  });
-};
+const styles = StyleSheet.create({
+  textInput,
+  // @ts-ignore
+  noOutline: isWebInternal() ? {outlineStyle: 'none'} : {},
+  textWrapFix: Platform.select({
+    ios: {
+      paddingVertical: 5,
+    },
+  }),
+});
