+++ template/src/components/GridVideo.tsx
@@ -12,7 +12,7 @@
 import {layoutComponent, useRtc} from 'customization-api';
 import React, {useMemo, useState} from 'react';
 import {View, StyleSheet, Dimensions, Pressable} from 'react-native';
-import {useIsWebInternal} from '../utils/common';
+import {isWebInternal} from '../utils/common';
 import {useSetPinnedLayout} from '../pages/video-call/DefaultLayouts';
 import RenderComponent from '../pages/video-call/RenderComponent';
 const layout = (len: number, isDesktop: boolean = true) => {
@@ -34,7 +34,6 @@ const layout = (len: number, isDesktop: boolean = true) => {
 };
 
 const GridVideo: layoutComponent = ({renderData}) => {
-  const isWebInternal = useIsWebInternal();
   const {dispatch} = useRtc();
   let onLayout = (e: any) => {
     setDim([
