+++ template/src/pages/Create.tsx
@@ -21,7 +21,7 @@ import Toast from '../../react-native-toast-message';
 import {ErrorContext} from '../components/common';
 import ShareLink from '../components/Share';
 import Logo from '../components/common/Logo';
-import {useIsWebInternal, isValidReactComponent} from '../utils/common';
+import {isWebInternal, isValidReactComponent} from '../utils/common';
 import {useCustomization} from 'customization-implementation';
 import {useString} from '../utils/useString';
 import useCreateMeeting from '../utils/useCreateMeeting';
@@ -32,7 +32,6 @@ import {MeetingInfoDefaultValue} from '../components/meeting-info/useMeetingInfo
 import {useSetMeetingInfo} from '../components/meeting-info/useSetMeetingInfo';
 
 const Create = () => {
-  const isWebInternal = useIsWebInternal();
   const {CreateComponent} = useCustomization((data) => {
     let components: {
       CreateComponent?: React.ElementType;
