+++ template/src/subComponents/NetworkQualityPill.tsx
@@ -13,7 +13,7 @@ import {
 import {useString} from '../utils/useString';
 import {networkIconsObject} from '../components/NetworkQualityContext';
 import {NetworkQualities} from 'src/language/default-labels/videoCallScreenLabels';
-import {useIsWebInternal} from '../utils/common';
+import {isWebInternal} from '../utils/common';
 import NetworkQualityContext from '../components/NetworkQualityContext';
 import {RenderInterface, UidType} from '../../agora-rn-uikit';
 
@@ -112,7 +112,6 @@ const PlatformSpecificWrapper = ({
   setNetworkTextVisible,
   children,
 }: any) => {
-  const isWebInternal = useIsWebInternal();
   return !isWebInternal() ? (
     <Pressable
       style={{
