+++ template/src/utils/isMobileOrTablet.native.ts
@@ -1,0 +1,5 @@
+const isMobileOrTablet = () => {
+  return true;
+};
+
+export default isMobileOrTablet;
