--- template/src/utils/useIsMobileOrTablet.native.ts
@@ -1,8 +1,0 @@
-const useIsMobileOrTablet = () => {
-  const isMobileOrTablet = () => {
-    return true;
-  };
-  return isMobileOrTablet;
-};
-
-export default useIsMobileOrTablet;
